split up code and add command line parsing
