+++ Cargo.toml
@@ -8,3 +8,4 @@ edition = "2018"
 
 [dependencies]
 image = "0.23.8"
+clap = "3.0.0-beta.1"
